Feature/consume stock (#45)
* Consume product from storeroom

In this commit we can consume some product quantity from a storeroom. When we consume a product quantity, the product stock is decreased, but if we want consume a quantity greater than product stock, we got an exception. Furthermore, when we consume a product, it raises an domain event.

* Bug solved in storeroom repository

When we got products from storeroom, the product's stock came to ZERO.